extract base writer
