+++ packages/dd-trace/src/ci-visibility/exporters/agentless/writer.js
@@ -3,9 +3,11 @@ const request = require('./request')
 const log = require('../../../log')
 
 const { AgentlessCiVisibilityEncoder } = require('../../../encode/agentless-ci-visibility')
+const BaseWriter = require('../../../exporters/common/writer')
 
-class Writer {
+class Writer extends BaseWriter {
   constructor ({ url, tags }) {
+    super(...arguments)
     const { 'runtime-id': runtimeId, env, service } = tags
     this._url = url
     this._encoder = new AgentlessCiVisibilityEncoder({ runtimeId, env, service })
@@ -17,8 +19,8 @@ class Writer {
     this._encoder.append(trace)
   }
 
-  _sendPayload (data, done) {
-    makeRequest(data, this._url, 15000, (err, res) => {
+  _sendPayload (data, _, done) {
+    makeRequest(data, this._url, (err, res) => {
       if (err) {
         log.error(err)
         done()
@@ -28,21 +30,9 @@ class Writer {
       done()
     })
   }
-
-  flush (done = () => {}) {
-    const count = this._encoder.count()
-
-    if (count > 0) {
-      const payload = this._encoder.makePayload()
-
-      this._sendPayload(payload, done)
-    } else {
-      done()
-    }
-  }
 }
 
-function makeRequest (data, url, timeout, cb) {
+function makeRequest (data, url, cb) {
   const options = {
     path: '/api/v2/citestcycle',
     method: 'POST',
@@ -50,7 +40,7 @@ function makeRequest (data, url, timeout, cb) {
       'Content-Type': 'application/msgpack',
       'dd-api-key': process.env.DATADOG_API_KEY || process.env.DD_API_KEY
     },
-    timeout
+    timeout: 15000
   }
 
   options.protocol = url.protocol
